+++ lib/engineyard-serverside/bundle_installer.rb
@@ -1,6 +1,0 @@
-module EY
-  module Serverside
-    class BundleInstaller < Struct.new(:version, :options)
-    end
-  end
-end